- Added new migrations for new initial database - Switched to scoped dependency injections

+++ backend/LuckyRest/Database/DTOs/Models/WorkshopMapDto.cs
@@ -7,7 +7,7 @@ public class WorkshopMapDto
     public string Name { get; set; } = string.Empty;
     public string ImageUrl { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
-    public int MapId { get; set; }
+    public long MapId { get; set; }
 
     public static WorkshopMapDto FromEntity(WorkshopMap? workshopMap)
     {
